Move setup logic into populateDB script, move db dump
